requirements.txt
streamlit
streamlit-feedback
google-generativeai
python-dotenv
langchain
langsmith
langchain_google_genai
langchain_core
langchain_community
regex
IPython
faiss-cpu

PyPDF2


-e .
